@@ -1,7 +1,8 @@
 #ifndef CONSTANTS_H
 #define CONSTANTS_H
 
-#define MAX_ROBOT_COUNT 12 //don't change
+#define MAX_ROBOT_COUNT 16 //don't change
+#define TEAM_COUNT 2
 
 #endif //CONSTANTS_H
 
