@@ -29,6 +29,7 @@ Copyright (C) 2011, Parsian Robotic Center (eew.aut.ac.ir/~parsian/grsim)
 
 #include <stdint.h>
 #include <stdio.h>
+#include <memory>
 
 #include <vartypes/VarTreeModel.h>
 #include <vartypes/VarItem.h>
@@ -69,22 +70,22 @@ using namespace VarTypes;
 #else
 
 #define DEF_VALUE(type,Type,name)  \
-            std::tr1::shared_ptr<VarTypes::Var##Type> v_##name; \
+            std::shared_ptr<VarTypes::Var##Type> v_##name; \
             inline type name() {return v_##name->get##Type();}
 
 #define DEF_FIELD_VALUE(type,Type,name)  \
-            std::tr1::shared_ptr<VarTypes::Var##Type> v_DivA_##name; \
-            std::tr1::shared_ptr<VarTypes::Var##Type> v_DivB_##name; \
+            std::shared_ptr<VarTypes::Var##Type> v_DivA_##name; \
+            std::shared_ptr<VarTypes::Var##Type> v_DivB_##name; \
             inline type name() {return (Division() == "Division A" ? v_DivA_##name: v_DivB_##name)->get##Type(); }
 
 #define DEF_ENUM(type,name)  \
-            std::tr1::shared_ptr<VarTypes::VarStringEnum> v_##name; \
+            std::shared_ptr<VarTypes::VarStringEnum> v_##name; \
             type name() {if(v_##name!=NULL) return v_##name->getString();return * (new type);}
 
 #define DEF_TREE(name)  \
-            std::tr1::shared_ptr<VarTypes::VarList> name;
+            std::shared_ptr<VarTypes::VarList> name;
 #define DEF_PTREE(parents, name)  \
-            std::tr1::shared_ptr<VarTypes::VarList> parents##_##name;
+            std::shared_ptr<VarTypes::VarList> parents##_##name;
 
 #endif
 
@@ -117,6 +118,14 @@ public:
     double WheelTangentFriction;
     double WheelPerpendicularFriction;
     double Wheel_Motor_FMax;
+    double MaxLinearKickSpeed;
+    double MaxChipKickSpeed;
+    double AccSpeedupAbsoluteMax;
+    double AccSpeedupAngularMax;
+    double AccBrakeAbsoluteMax;
+    double AccBrakeAngularMax;
+    double VelAbsoluteMax;
+    double VelAngularMax;
 };
 
 
@@ -130,12 +139,12 @@ protected:
 public:
   VarListPtr geo_vars;
   ConfigWidget();
-  virtual ~ConfigWidget();
+  ~ConfigWidget() override;
 
   QSettings* robot_settings;
-  RobotSettings robotSettings;
-  RobotSettings blueSettings;
-  RobotSettings yellowSettings;
+  RobotSettings robotSettings{};
+  RobotSettings blueSettings{};
+  RobotSettings yellowSettings{};
 
   /*    Geometry/Game Vartypes   */
   DEF_ENUM(std::string, Division)
@@ -155,6 +164,9 @@ public:
   DEF_FIELD_VALUE(double,Double,Goal_Depth)
   DEF_FIELD_VALUE(double,Double,Goal_Width)
   DEF_FIELD_VALUE(double,Double,Goal_Height)
+  DEF_VALUE(double,Double,Camera_Height)
+  DEF_VALUE(double,Double,Camera_Focal_Length)
+  DEF_VALUE(double,Double,Camera_Scaling_Limit)
 
   DEF_ENUM(std::string,YellowTeam)
   DEF_ENUM(std::string,BlueTeam)
@@ -166,17 +178,28 @@ public:
   DEF_VALUE(double,Double,BallBounceVel)
   DEF_VALUE(double,Double,BallLinearDamp)
   DEF_VALUE(double,Double,BallAngularDamp)
+  DEF_VALUE(bool,Bool,BallProjectAirborne)
+  DEF_VALUE(double,Double,BallModelTwoPhaseAccSlide)
+  DEF_VALUE(double,Double,BallModelTwoPhaseAccRoll)
+  DEF_VALUE(double,Double,BallModelTwoPhaseKSwitch)
+  DEF_VALUE(double,Double,BallModelChipFixedLossDampingXyFirstHop)
+  DEF_VALUE(double,Double,BallModelChipFixedLossDampingXyOtherHops)
+  DEF_VALUE(double,Double,BallModelChipFixedLossDampingZ)
 
   DEF_VALUE(bool,Bool,SyncWithGL)
   DEF_VALUE(double,Double,DesiredFPS)
   DEF_VALUE(double,Double,DeltaTime)
   DEF_VALUE(int,Int,sendGeometryEvery)
   DEF_VALUE(double,Double,Gravity)
+  DEF_VALUE(bool,Bool,ResetTurnOver)
   DEF_VALUE(std::string,String,VisionMulticastAddr)  
   DEF_VALUE(int,Int,VisionMulticastPort)  
   DEF_VALUE(int,Int,CommandListenPort)
   DEF_VALUE(int,Int,BlueStatusSendPort)
   DEF_VALUE(int,Int,YellowStatusSendPort)
+  DEF_VALUE(int,Int,SimControlListenPort)
+  DEF_VALUE(int,Int,BlueControlListenPort)
+  DEF_VALUE(int,Int,YellowControlListenPort)
   DEF_VALUE(int,Int,sendDelay)
   DEF_VALUE(bool,Bool,noise)
   DEF_VALUE(double,Double,noiseDeviation_x)
@@ -188,7 +211,11 @@ public:
   DEF_VALUE(double,Double,yellow_team_vanishing)
   DEF_VALUE(std::string, String, plotter_addr)
   DEF_VALUE(int, Int, plotter_port)
-  DEF_VALUE(bool, Bool, plotter)  
+  DEF_VALUE(bool, Bool, plotter)
+
+  DEF_VALUE(std::string, String, ColorRobotBlue)
+  DEF_VALUE(std::string, String, ColorRobotYellow)
+
   void loadRobotSettings(QString team);
 public slots:  
   void loadRobotsSettings();
