@@ -24,9 +24,9 @@
 #include <string>
 #include <QMutex>
 #include <QObject>
-#include "messages_robocup_ssl_detection.pb.h"
-#include "messages_robocup_ssl_geometry.pb.h"
-#include "messages_robocup_ssl_wrapper.pb.h"
+#include "ssl_vision_detection.pb.h"
+#include "ssl_vision_geometry.pb.h"
+#include "ssl_vision_wrapper.pb.h"
 using namespace std;
 
 class QUdpSocket;
