@@ -40,8 +40,8 @@ using namespace std;
 
 #define BS_TO_OMPL 0.1
 #define OMPL_TO_BS 10
-#define HALF_FIELD_MAXX 4500
-#define HALF_FIELD_MAXY 3000
+#define HALF_FIELD_MAXX 6000
+#define HALF_FIELD_MAXY 4500
 #define HALF_FIELD_MAXX_OMPL HALF_FIELD_MAXX*BS_TO_OMPL
 #define HALF_FIELD_MAXY_OMPL HALF_FIELD_MAXY*BS_TO_OMPL
 #define self_radius 270*BS_TO_OMPL
