@@ -13,17 +13,17 @@ BS_ADDRESS                   = ['127.0.0.1:11211']
 GOAL_DEPTH                   = SELECT(300, 300)
 CENTER_X                     = SELECT(0, 0)
 CENTER_Y                     = SELECT(0, 0)
-HALF_FIELD_MAXX              = SELECT(4500, 4500)
-HALF_FIELD_MAXY              = SELECT(3000, 3000)
-OUR_GOAL_MAXY                = SELECT(350,	500)
-OUR_GOAL_MINY                = SELECT(-350, -500)
+HALF_FIELD_MAXX              = SELECT(6000, 4500)
+HALF_FIELD_MAXY              = SELECT(4500, 3000)
+OUR_GOAL_MAXY                = SELECT(600,	500)
+OUR_GOAL_MINY                = SELECT(-600, -500)
 OUR_GOAL_WIDTH               = OUR_GOAL_MAXY - OUR_GOAL_MINY
 CENTER_CIRCLE_DIAMETER       = SELECT(1000,1000)
 CENTER_CIRCLE_RADIUS		 = SELECT(500,500)
-DBOX_WIDTH                   = SELECT(1400,2000)	#Along y       
-DBOX_HEIGHT					 = SELECT(700,1000)	    #Along x      
-OUR_DBOX_MAXY				 = SELECT(700,1000)
-OUR_DBOX_MINY				 = SELECT(-700,-1000)
+DBOX_WIDTH                   = SELECT(2400,2000)	#Along y       
+DBOX_HEIGHT					 = SELECT(1200,1000)	    #Along x      
+OUR_DBOX_MAXY				 = SELECT(1200,1000)
+OUR_DBOX_MINY				 = SELECT(-1200,-1000)
 OUR_DBOX_X 					 =-HALF_FIELD_MAXX + DBOX_HEIGHT
 # Old DBOX Constants
 DBOX_WIDTH_OLD                   = SELECT(1000,1000)       # Along X direction
